requirements.txt
pymupdf
sentence-transformers
faiss-cpu
numpy
scikit-learn
tqdm
spacy
google-generativeai
python-dotenv